requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.34
pydantic>=2.10.0
python-dotenv
langchain-ollama
gtts
